refactor: use sitelocale from datocms types directly
