+++ src/lib/seo/index.ts
@@ -1,6 +1,5 @@
-import type { Tag } from '@lib/datocms/types';
+import type { Tag, SiteLocale } from '@lib/datocms/types';
 import { getLocale } from '@lib/i18n';
-import type { SiteLocale } from '@lib/i18n/types';
 import { globalSeo } from '@lib/site.json';
 import aiRobotsTxt from './ai.robots.txt?raw';
