+++ src/middleware/i18n.ts
@@ -1,6 +1,5 @@
 import { defineMiddleware } from 'astro:middleware';
-import { defaultLocale, locales, setLocale } from '@lib/i18n';
-import type { SiteLocale } from '@lib/i18n/types';
+import { defaultLocale, isLocale, setLocale } from '@lib/i18n';
 
 /**
  * i18n middleware:
@@ -11,12 +10,12 @@ export const i18n = defineMiddleware(async ({ params, request }, next) => {
     // if the locale param is unavailable, it didn't match a [locale]/* route
     // so we attempt to extract the locale from the URL and fallback to the default locale
     const pathLocale = new URL(request.url).pathname.split('/')[1];
-    const locale = locales.includes(pathLocale as SiteLocale)
+    const locale = isLocale(pathLocale)
       ? pathLocale
       : defaultLocale;
     Object.assign(params, { locale });
   }
-  setLocale(params.locale as SiteLocale);
+  setLocale(params.locale);
 
   const response = await next();
   return response;
