+++ src/pages/404.astro
@@ -1,6 +1,5 @@
 ---
-import type { SiteLocale } from '@lib/i18n/types';
-import type { NotFoundPageQuery } from '@lib/datocms/types';
+import type { NotFoundPageQuery, SiteLocale } from '@lib/datocms/types';
 import { datocmsRequest } from '@lib/datocms';
 import { noIndexTag, titleTag } from '@lib/seo';
 import Layout from '@layouts/Default.astro';
