+++ src/pages/[locale]/search/results.partial.astro
@@ -1,7 +1,7 @@
 ---
 import { hasValidQuery, queryParamName } from '@lib/search';
 import { datocmsSearch } from '@lib/datocms';
-import type { SiteLocale } from '@lib/i18n/types';
+import type { SiteLocale } from '@lib/datocms/types';
 import SearchResult from '@components/SearchResult.astro';
 
 export const partial = true;
